Move cert_expiry coordinator to its own file
