+++ homeassistant/components/cert_expiry/__init__.py
@@ -1,22 +1,13 @@
 """The cert_expiry component."""
 from __future__ import annotations
 
-from datetime import datetime, timedelta
-import logging
-
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_HOST, CONF_PORT, Platform
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.start import async_at_started
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 
-from .const import DEFAULT_PORT, DOMAIN
-from .errors import TemporaryFailure, ValidationFailure
-from .helper import get_cert_expiry_timestamp
-
-_LOGGER = logging.getLogger(__name__)
-
-SCAN_INTERVAL = timedelta(hours=12)
+from .const import DOMAIN
+from .coordinator import CertExpiryDataUpdateCoordinator
 
 PLATFORMS = [Platform.SENSOR]
 
@@ -45,37 +36,3 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
 async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Unload a config entry."""
     return await hass.config_entries.async_unload_platforms(entry, PLATFORMS)
-
-
-class CertExpiryDataUpdateCoordinator(DataUpdateCoordinator[datetime | None]):
-    """Class to manage fetching Cert Expiry data from single endpoint."""
-
-    def __init__(self, hass, host, port):
-        """Initialize global Cert Expiry data updater."""
-        self.host = host
-        self.port = port
-        self.cert_error = None
-        self.is_cert_valid = False
-
-        display_port = f":{port}" if port != DEFAULT_PORT else ""
-        name = f"{self.host}{display_port}"
-
-        super().__init__(
-            hass, _LOGGER, name=name, update_interval=SCAN_INTERVAL, always_update=False
-        )
-
-    async def _async_update_data(self) -> datetime | None:
-        """Fetch certificate."""
-        try:
-            timestamp = await get_cert_expiry_timestamp(self.hass, self.host, self.port)
-        except TemporaryFailure as err:
-            raise UpdateFailed(err.args[0]) from err
-        except ValidationFailure as err:
-            self.cert_error = err
-            self.is_cert_valid = False
-            _LOGGER.error("Certificate validation error: %s [%s]", self.host, err)
-            return None
-
-        self.cert_error = None
-        self.is_cert_valid = True
-        return timestamp
