+++ tests/components/cert_expiry/test_init.py
@@ -42,7 +42,7 @@ async def test_setup_with_config(hass: HomeAssistant) -> None:
     with patch(
         "homeassistant.components.cert_expiry.config_flow.get_cert_expiry_timestamp"
     ), patch(
-        "homeassistant.components.cert_expiry.get_cert_expiry_timestamp",
+        "homeassistant.components.cert_expiry.coordinator.get_cert_expiry_timestamp",
         return_value=future_timestamp(1),
     ):
         await hass.async_block_till_done()
@@ -63,7 +63,7 @@ async def test_update_unique_id(hass: HomeAssistant) -> None:
     assert not entry.unique_id
 
     with patch(
-        "homeassistant.components.cert_expiry.get_cert_expiry_timestamp",
+        "homeassistant.components.cert_expiry.coordinator.get_cert_expiry_timestamp",
         return_value=future_timestamp(1),
     ):
         assert await async_setup_component(hass, DOMAIN, {}) is True
@@ -91,7 +91,7 @@ async def test_unload_config_entry(mock_now, hass: HomeAssistant) -> None:
 
     timestamp = future_timestamp(100)
     with patch(
-        "homeassistant.components.cert_expiry.get_cert_expiry_timestamp",
+        "homeassistant.components.cert_expiry.coordinator.get_cert_expiry_timestamp",
         return_value=timestamp,
     ):
         assert await async_setup_component(hass, DOMAIN, {}) is True
@@ -134,7 +134,7 @@ async def test_delay_load_during_startup(hass: HomeAssistant) -> None:
 
     timestamp = future_timestamp(100)
     with patch(
-        "homeassistant.components.cert_expiry.get_cert_expiry_timestamp",
+        "homeassistant.components.cert_expiry.coordinator.get_cert_expiry_timestamp",
         return_value=timestamp,
     ):
         await hass.async_start()
