+++ tests/components/cert_expiry/test_sensors.py
@@ -29,7 +29,7 @@ async def test_async_setup_entry(mock_now, hass: HomeAssistant) -> None:
     timestamp = future_timestamp(100)
 
     with patch(
-        "homeassistant.components.cert_expiry.get_cert_expiry_timestamp",
+        "homeassistant.components.cert_expiry.coordinator.get_cert_expiry_timestamp",
         return_value=timestamp,
     ):
         entry.add_to_hass(hass)
@@ -83,7 +83,7 @@ async def test_update_sensor(hass: HomeAssistant) -> None:
     timestamp = future_timestamp(100)
 
     with patch("homeassistant.util.dt.utcnow", return_value=starting_time), patch(
-        "homeassistant.components.cert_expiry.get_cert_expiry_timestamp",
+        "homeassistant.components.cert_expiry.coordinator.get_cert_expiry_timestamp",
         return_value=timestamp,
     ):
         entry.add_to_hass(hass)
@@ -99,7 +99,7 @@ async def test_update_sensor(hass: HomeAssistant) -> None:
 
     next_update = starting_time + timedelta(hours=24)
     with patch("homeassistant.util.dt.utcnow", return_value=next_update), patch(
-        "homeassistant.components.cert_expiry.get_cert_expiry_timestamp",
+        "homeassistant.components.cert_expiry.coordinator.get_cert_expiry_timestamp",
         return_value=timestamp,
     ):
         async_fire_time_changed(hass, utcnow() + timedelta(hours=24))
@@ -127,7 +127,7 @@ async def test_update_sensor_network_errors(hass: HomeAssistant) -> None:
     timestamp = future_timestamp(100)
 
     with patch("homeassistant.util.dt.utcnow", return_value=starting_time), patch(
-        "homeassistant.components.cert_expiry.get_cert_expiry_timestamp",
+        "homeassistant.components.cert_expiry.coordinator.get_cert_expiry_timestamp",
         return_value=timestamp,
     ):
         entry.add_to_hass(hass)
@@ -156,7 +156,7 @@ async def test_update_sensor_network_errors(hass: HomeAssistant) -> None:
     assert state.state == STATE_UNAVAILABLE
 
     with patch("homeassistant.util.dt.utcnow", return_value=next_update), patch(
-        "homeassistant.components.cert_expiry.get_cert_expiry_timestamp",
+        "homeassistant.components.cert_expiry.coordinator.get_cert_expiry_timestamp",
         return_value=timestamp,
     ):
         async_fire_time_changed(hass, utcnow() + timedelta(hours=48))
